+++ snow-mailbox-client/src/components/mailboxImg.tsx
